Extract pure helpers into lib.js
Move classifyCommand, resolveSocketWaitMs, shellQuote, LldbError, and
the SOCKET_WAIT_MS_* constants out of index.js into a new lib.js. The
goal is testability: these are the functions most prone to silent
regression (command-classification edge cases, env-var fallback
logic), and importing index.js starts the MCP server via top-level
await, so they can't be unit-tested in place.

lib.js has no side effects on import. resolveSocketWaitMs gains a
second, defaulted `env = process.env` argument so tests can pass a
synthetic env; production callers are unchanged.
